Refactoring RunningSpeed related code to pass both environments
