+++ allensdk/brain_observatory/ecephys/ecephys_api/ecephys_api.py
@@ -3,7 +3,7 @@ from typing import Dict
 import numpy as np
 import pandas as pd
 
-from ... import RunningSpeed
+from ...running_speed import RunningSpeed
 
 class EcephysApi:
